newsfetch-examples/requirements.txt
python_dotenv~=0.20.0
opensearch-py
sentence-transformers
